# openapi/Penguin.yaml
openapi: 3.0.3
info:
  version: 0.6.0
  title: Penguin API
  description: "A list of the currently provided Interfaces to interact with the penguin, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local pen node main API
      port:
        default: "1633"
        description: Service port provided in pen node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local pen node main API
      port:
        default: "1633"
        description: Service port provided in pen node config
paths:
...
  "/feeds/{owner}/{topic}":
    post:
      summary: Create an initial feed root manifest
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: type
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "401":
          $ref: "PenguinCommon.yaml#/components/responses/401"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: Find feed update
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: at
          schema:
            type: integer
          required: false
          description: "Timestamp of the update (default: now)"
        - in: query
          name: type
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
      responses:
        "200":
          description: Latest feed update
          headers:
            "penguin-feed-index":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinFeedIndex"
            "penguin-feed-index-next":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinFeedIndexNext"
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "401":
          $ref: "PenguinCommon.yaml#/components/responses/401"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
...